Minor refactoring in Microsoft.Azure.IIoT.Services.Processor.Events.
